- Remove binding.pry from create_purchase_order.rb file. - Add code for assign SO number in create_sales_order_in_sap.rb. - Add some parameters in SO payload for creating Sales Order in SAP with this order number which is assign in sprint.

+++ app/services/services/overseers/sales_orders/create_sales_order_in_sap.rb
@@ -6,17 +6,23 @@ class Services::Overseers::SalesOrders::CreateSalesOrderInSap < Services::Shared
 
   def call
     if params['approve'].present?
-      doc_id = ::Resources::Order.create(sales_order)
-      order = ::Resources::Order.find(doc_id)
-      sales_order.update_attributes(remote_status: :'Supplier PO: Request Pending', status: :'Approved', mis_date: Date.today, order_number: order['DocNum'], remote_uid: order['DocEntry'])
-      Services::Overseers::Inquiries::UpdateStatus.new(sales_order, :order_won).call
-      comment = sales_order.inquiry.comments.create!(message: 'SAP Approved', overseer: Overseer.default_approver, sales_order: sales_order)
-      if sales_order.approval.blank?
-        sales_order.create_approval!(comment: comment, overseer: Overseer.default_approver)
-        sales_order.rejection.destroy! if sales_order.rejection.present?
+      series_name = sales_order.sales_quote.bill_from.series_code + ' ' + Date.today.year.to_s
+      series = Series.where(document_type: 'Sales Order', series_name: series_name)
+      if series.present?
+        sales_order.update_attributes(remote_status: :'Supplier PO: Request Pending', status: :'Approved', mis_date: Date.today, order_number: series.first.last_number)
+        series.first.increment_last_number
+        doc_id = ::Resources::Order.create(sales_order)
+        order = ::Resources::Order.find(doc_id)
+        sales_order.update_attributes(remote_uid: order['DocEntry'])
+        Services::Overseers::Inquiries::UpdateStatus.new(sales_order, :order_won).call
+        comment = sales_order.inquiry.comments.create!(message: 'SAP Approved', overseer: Overseer.default_approver, sales_order: sales_order)
+        if sales_order.approval.blank?
+          sales_order.create_approval!(comment: comment, overseer: Overseer.default_approver)
+          sales_order.rejection.destroy! if sales_order.rejection.present?
+        end
+        # sales_order.serialized_pdf.attach(io: File.open(RenderPdfToFile.for(sales_order)), filename: sales_order.filename)
+        sales_order.update_index
       end
-      # sales_order.serialized_pdf.attach(io: File.open(RenderPdfToFile.for(sales_order)), filename: sales_order.filename)
-      sales_order.update_index
     else
       sales_order.update_attributes(status: :'SAP Rejected')
       comment = InquiryComment.where(message: "Rejected by Accounts. Reason: #{params[:sales_order][:custom_fields][:message]}",
